+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Refined Storage Changelog
 
 ### 1.5.14
+- Updated Forge to 2426 (raoulvdberge)
+- Added Albedo support (raoulvdberge)
 - Fixed more crashes relating to scrollbar in GUIs (raoulvdberge)
 - Fixed crash with Detector (raoulvdberge)
 - Fixed bug where pattern create button wasn't visible when grid tabs were selected (raoulvdberge)